fix: Convert all confidence scores from Double to BigDecimal with scale 2
**Problem**: Floating point arithmetic causes precision errors like 0.30 != 0.300000000000000001

**Solution**: Use BigDecimal with scale 2 and RoundingMode.HALF_UP for all confidence scores

**Changes**:

Domain Entities:
- FundingSourceCandidate: confidenceScore Double → BigDecimal
- Domain: bestConfidenceScore Double → BigDecimal
- DiscoverySession: averageConfidenceScore Double → BigDecimal
- EnhancementRecord: confidenceScore Double → BigDecimal

Database:
- V8__create_domain.sql: DOUBLE PRECISION → DECIMAL(3,2) with CHECK constraint
- V1, V4, V5 migrations: Already using DECIMAL(3,2)

Service Layer:
- CandidateValidationService: Update threshold comparison to use BigDecimal.compareTo()
- DiscoveryOrchestrationService: Rewrite calculateAverageConfidence() with BigDecimal arithmetic
- MetadataJudgingService: Complete BigDecimal implementation with scale 2, RoundingMode.HALF_UP
- DomainRegistryService: Update method signatures

Repositories:
- FundingSourceCandidateRepository: findByConfidenceScoreGreaterThan() parameter Double → BigDecimal

Test Files (All Updated):
- Service tests: CandidateValidationServiceTest, DiscoveryOrchestrationServiceTest
- Repository integration tests: DomainRepositoryIT, DiscoverySessionRepositoryIT, EnhancementRecordRepositoryIT, FundingSourceCandidateRepositoryIT
- Integration tests: AIEnhancementIntegrationTest, AuditTrailIntegrationTest, DiscoveryWorkflowIntegrationTest, DuplicateDetectionIntegrationTest
- TestDataFactory: Update all averageConfidenceScore builders
- All test builders: .confidenceScore(double) → .confidenceScore(new BigDecimal("X.XX"))
- All test assertions: .isEqualTo(double) → .isEqualByComparingTo(new BigDecimal("X.XX"))
- All test comparisons: >= double → .compareTo(new BigDecimal("X.XX")) >= 0

**Build Status**: ✅ Compilation succeeds (mvn clean compile)

**Constitution Compliance**: Updated to v1.4.0 with "Data Precision Standards (CRITICAL)" section

+++ backend/src/main/java/com/northstar/funding/discovery/domain/DiscoverySession.java
@@ -1,5 +1,6 @@
 package com.northstar.funding.discovery.domain;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -82,8 +83,12 @@ public class DiscoverySession {
     @Column("sources_scraped")
     private Integer sourcesScraped;
     
+    /**
+     * Average AI confidence score of candidates found (0.00-1.00)
+     * Uses BigDecimal with scale 2 for precise decimal arithmetic
+     */
     @Column("average_confidence_score")
-    private Double averageConfidenceScore;
+    private BigDecimal averageConfidenceScore;
     
     // Error Handling & Diagnostics
     @Column("error_messages")
